+++ index.html
@@ -224,14 +224,7 @@
 </body>
 <script src="http://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script>window.jQuery || document.write('<script src="jquery.min.js"><\/script>')</script>
-<script type="text/javascript" src="js/data.js"></script>
-<script type="text/javascript" src="js/DOM.js"></script>
-<script type="text/javascript" src="js/Puzzle.js"></script>
-<script type="text/javascript" src="js/Player.js"></script>
-<script type="text/javascript" src="js/Wheel.js"></script>
-<script type="text/javascript" src="js/Round.js"></script>
-<script type="text/javascript" src="js/Game.js"></script>
-<script type="text/javascript" src="js/index.js"></script>
+<script type="module" src="js/index.js"></script>
 </html>
