+++ js/Player.js
@@ -1,3 +1,10 @@
+import data from './data.js';
+import Game from './Game.js';
+import domUpdates from './DOM.js';
+import Puzzle from './Puzzle.js';
+import Round from './Round.js';
+import Wheel from './Wheel.js';
+
 class Player {
   constructor(name) {
     this.name = name;
@@ -19,6 +26,4 @@ class Player {
 }
 
 
-if (typeof module !== 'undefined') {
-  module.exports = Player;
-}
+export default Player;
