+++ js/data.js
@@ -817,6 +817,4 @@ const data = {
 };
 
 
-if (typeof module !== 'undefined') {
-  module.exports = data;
-}
+export default data;
